+++ tests/unit/utils/validation.tests.ts
@@ -42,6 +42,12 @@ test('Settings Validation', () => {
         lightColorScheme: '',
         darkColorScheme: false,
         immediateModify: 1,
+        darkSchemeCssVariableBg: '',
+        lightSchemeCssVariableBg: '',
+        darkSchemeCssVariableText: '',
+        lightSchemeCssVariableText: '',
+        cssVariableScrollBar: '',
+        cssVariableSelection: '',
     };
     themeValidation = validateTheme(wonkyTheme as any);
     expect(themeValidation.errors.length).toBeGreaterThan(0);
@@ -76,6 +82,12 @@ test('Settings Validation', () => {
             lightColorScheme: '',
             darkColorScheme: false,
             immediateModify: 1,
+            darkSchemeCssVariableBg: '',
+            lightSchemeCssVariableBg: '',
+            darkSchemeCssVariableText: '',
+            lightSchemeCssVariableText: '',
+            cssVariableScrollBar: '',
+            cssVariableSelection: '',
         },
         presets: [
             {id: '', name: 'P1', urls: ['a.com'], theme: {brightness: 100}},
@@ -171,6 +183,12 @@ test('Settings Validation', () => {
             lightColorScheme: 'Lightness',
             darkColorScheme: 'Darkness',
             immediateModify: true,
+            darkSchemeCssVariableBg: '--background',
+            lightSchemeCssVariableBg: '--foreground',
+            darkSchemeCssVariableText: '--color0',
+            lightSchemeCssVariableText: '--color1',
+            cssVariableScrollBar: '--color2',
+            cssVariableSelection: '--color3',
         },
         presets: [
             {id: 'p5', name: 'P5', urls: ['a.com'], theme: {brightness: 100} as Theme},